gem5: reduced generation id to 8 bits.

+++ src/libs/rust/base/src/tcu/mod.rs
@@ -55,7 +55,7 @@ pub type Label = u64;
 /// A activity id
 pub type ActId = u16;
 /// A tile-generation id
-pub type GenId = u16;
+pub type GenId = u8;
 
 #[cfg(M3_TARGET = "gem5")]
 pub const EXREG_REGS: usize = 16;
